requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mysql-connector-python==9.3.0
namex==0.1.0
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
protobuf==5.29.5
Pygments==2.19.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
